requirements.txt
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
openai>=1.10.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
python-dotenv>=1.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
reportlab>=4.0.0
